Add OCPI support & NDW Netherlands source

+++ evmap_backend/sync.py
@@ -2,6 +2,7 @@ from typing import Iterable, List, Tuple
 
 from django.db import transaction
 from django.forms import model_to_dict
+from tqdm import tqdm
 
 from evmap_backend.chargers.models import Chargepoint, ChargingSite, Connector
 
@@ -17,7 +18,7 @@ def sync_chargers(
                 "id", flat=True
             )
         )
-        for site, chargepoints in sites:
+        for site, chargepoints in tqdm(sites):
             site, created = ChargingSite.objects.update_or_create(
                 data_source=data_source,
                 id_from_source=site.id_from_source,
@@ -32,7 +33,7 @@ def sync_chargers(
                     site_ids_to_delete.remove(site.id)
                 except KeyError:
                     raise ValueError(
-                        f"ID {site.id_from_source} seems to appear more than once in DATEX2 data!"
+                        f"ID {site.id_from_source} seems to appear more than once in input data!"
                     )
 
             chargepoint_ids_to_delete = set(
@@ -50,26 +51,51 @@ def sync_chargers(
                 if not created:
                     chargepoint_ids_to_delete.remove(chargepoint.id)
 
-                    # check if the existing connectors match the updated ones
-                    new_connectors = [
-                        model_to_dict(connector, exclude=["chargepoint", "id"])
-                        for connector in connectors
-                    ]
-                    existing_connectors = [
-                        model_to_dict(conn, exclude=["chargepoint", "id"])
-                        for conn in chargepoint.connectors.all()
-                    ]
-                    if len(new_connectors) == len(existing_connectors) and all(
-                        conn in existing_connectors for conn in new_connectors
-                    ):
-                        continue
+                if all(conn.id_from_source != "" for conn in connectors):
+                    connector_ids_to_delete = set(
+                        Connector.objects.filter(chargepoint=chargepoint).values_list(
+                            "id", flat=True
+                        )
+                    )
+                    for connector in connectors:
+                        connector, created = Connector.objects.update_or_create(
+                            chargepoint=chargepoint,
+                            id_from_source=connector.id_from_source,
+                            defaults=model_to_dict(
+                                connector,
+                                exclude=["chargepoint", "id_from_source", "id"],
+                            ),
+                        )
+                        if not created:
+                            connector_ids_to_delete.remove(connector.id)
 
-                # delete all connectors and save the new ones to the DB
-                chargepoint.connectors.all().delete()
+                        # delete missing connectors
+                        Connector.objects.filter(
+                            id__in=connector_ids_to_delete
+                        ).delete()
+                else:
+                    # connector IDs are not available -> fallback
+                    if not created:
+                        # check if the existing connectors match the updated ones
+                        new_connectors = [
+                            model_to_dict(connector, exclude=["chargepoint", "id"])
+                            for connector in connectors
+                        ]
+                        existing_connectors = [
+                            model_to_dict(conn, exclude=["chargepoint", "id"])
+                            for conn in chargepoint.connectors.all()
+                        ]
+                        if len(new_connectors) == len(existing_connectors) and all(
+                            conn in existing_connectors for conn in new_connectors
+                        ):
+                            continue
+                        else:
+                            # delete all connectors and save the new ones to the DB
+                            chargepoint.connectors.all().delete()
 
-                for connector in connectors:
-                    connector.chargepoint = chargepoint
-                    connector.save()
+                    for connector in connectors:
+                        connector.chargepoint = chargepoint
+                        connector.save()
 
             # delete missing chargepoints
             Chargepoint.objects.filter(id__in=chargepoint_ids_to_delete).delete()
